The two remotes diverged: Gitea carried the aislop complexity refactor
(#33), the flaky StopWatch test fix (#31), and a CI docs fix, while GitHub
carried a fleet-standard .editorconfig/DotSettings sync, a whole-repo
jb cleanupcode sweep, and reasons added to inline aislop suppressions.
Neither side could fast-forward, so the mirror had been stuck since
2026-08-15.

All ten conflicts were the same shape: Gitea refactored code that GitHub
reformatted in place. Gitea's side won every one, because in each case the
refactor supersedes what GitHub's version documents or reformats:

- UsnJournalEntry moved to an options object and a JournalEntryFactory test
  helper, which removes the too-many-params problem that GitHub's new
  suppression comment merely explained. The suppression is dropped as moot.
- BrokerFrame.cs and JournalBrokerScanSession.Start.cs extracted types and
  code out of BrokerProtocol.cs and JournalBrokerScanSession.cs; taking
  GitHub's reformatted in-place copies would have duplicated the types.
- FreeUsnJournalInfo/FreeUsnJournalResult moved to unique_ptr RAII, so the
  manual deletes and their cpp-manual-delete suppressions are both gone.

GitHub's non-conflicting hunks are preserved, including the redundant-cast
removal in usn_journal.cpp: newCapacity and entryCount are both uint64_t
and sizeof yields size_t, so the casts were no-ops on this x64-only target.

This is a pure reconciliation: every file is either Gitea's content or
GitHub's, with no third-party reformatting mixed in.

Verified on the Linux lane (scripts/coverage-linux.sh), which is the same
job CI runs: 8 native smoke tests and 281 managed tests pass, 34 skipped,
315 total, identical to the pre-merge baseline. The Windows lane cannot run
here and is left to CI.

+++ MFTLibNative/mft/mft.internal.h
@@ -12,7 +12,6 @@
 #include "../internal.h"
 
 namespace mftlib::ntfs {
-
 struct DataRun {
     int64_t clusterOffset;
     uint64_t clusterCount;
@@ -39,16 +38,15 @@ uint8_t* ReadNonResidentData(HANDLE volumeHandle, PATTRIBUTE_RECORD_HEADER attr,
 bool ReadMFTRecord(HANDLE volumeHandle, const std::vector<DataRun>& mftRuns, uint32_t bytesPerCluster, uint8_t* buffer,
                    uint64_t recordNumber);
 #endif
-}  // namespace detail
-
-}  // namespace mftlib::ntfs
+} // namespace detail
+} // namespace mftlib::ntfs
 
 // Bundles the filename-filter parameters so they travel as a single argument
 // (and cannot be transposed) through the record-parsing path.
 struct FilterSpec {
-    const wchar_t* text;  // null = no filter (accept every named record)
-    uint16_t length;      // wchar_t units in text
-    uint32_t flags;       // match bitfield: 1=exact, 2=substring, 4=resolve paths
+    const wchar_t* text; // null = no filter (accept every named record)
+    uint16_t length; // wchar_t units in text
+    uint32_t flags; // match bitfield: 1=exact, 2=substring, 4=resolve paths
 };
 
 // Half-open record range [start, end) within a chunk buffer.
@@ -72,14 +70,14 @@ struct PathLookup {
     std::atomic<uint64_t> namesDropped;
 
     bool init(uint64_t totalRecords) {
-        parents = static_cast<uint64_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint64_t)));
-        nameLens = static_cast<uint8_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint8_t)));
-        nameOffsets = static_cast<uint32_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint32_t)));
+        parents = static_cast<uint64_t*>(calloc(totalRecords, sizeof(uint64_t)));
+        nameLens = static_cast<uint8_t*>(calloc(totalRecords, sizeof(uint8_t)));
+        nameOffsets = static_cast<uint32_t*>(calloc(totalRecords, sizeof(uint32_t)));
         // Each name entry can be up to 255 WCHAR units = 510 bytes; use 32 bytes avg * 2 for bytes.
         // A test hook can shrink the pool to exercise the exhaustion path.
         uint64_t capacityOverride = NamePoolCapacityOverride();
-        namePoolCapacity = (capacityOverride != 0U) ? capacityOverride : totalRecords * 64;  // bytes
-        namePool = static_cast<uint8_t*>(malloc(static_cast<size_t>(namePoolCapacity)));
+        namePoolCapacity = (capacityOverride != 0U) ? capacityOverride : totalRecords * 64; // bytes
+        namePool = static_cast<uint8_t*>(malloc(namePoolCapacity));
         namePoolUsed = 0;
         namesDropped = 0;
         return (parents != nullptr) && (nameLens != nullptr) && (nameOffsets != nullptr) && (namePool != nullptr);
@@ -95,7 +93,7 @@ struct PathLookup {
             return;
         }
         nameOffsets[recordIndex] = static_cast<uint32_t>(offset);
-        memcpy(namePool + offset, name, static_cast<size_t>(byteCount));
+        memcpy(namePool + offset, name, byteCount);
         nameLens[recordIndex] = nameLen;
     }
 
@@ -112,7 +110,7 @@ struct PathLookup {
 struct SliceResult {
     std::vector<MftFileEntry> entries;
 
-    void init(uint64_t cap) { entries.reserve(static_cast<size_t>(cap)); }
+    void init(uint64_t cap) { entries.reserve(cap); }
 };
 
 // Read-only inputs that steer record scanning: the name filter, the optional
